A lot has changed very sloppy code as I was rushing to get the scraper online. That said, I change all the code to run once on a evulate function in the handleStart function to greatly reduce time it takes.
Also commented out the finical data as this will slow down the script a bit too.

I also have is setup for local development at the moment. SO I will run it localy then I will add it to cloud.

 function $x(path) {
  	let xpathResult = document.evaluate(path, document, null, XPathResult.ORDERED_NODE_ITERATOR_TYPE, null);
     result = [];
   while (elem = xpathResult.iterateNext()) {
      result.push(elem);
   }

   $result = jQuery([]).pushStack( result );
   return $result;
}

The function above is very important as $x is only in the console and in puppeteer not shipped with jquery super weird.

Furthermore you can have low dept with the flatten packages and by merging objects.

with obj..., ...obj2

Handy stuff speak soon